+++ index.js
@@ -936,6 +936,7 @@
        *
        * @param  {String} name
        * @param  {Object} params
+       * @param  {Object} newContext
        * @param  {Boolean} keepURL
        *
        * @return {Promise<Boolean>}
@@ -943,7 +944,7 @@
 
     }, {
       key: "substitute",
-      value: function substitute(name, params, keepURL) {
+      value: function substitute(name, params, newContext, keepURL) {
         var _this3 = this;
 
         if (process.env.NODE_ENV !== 'production') {
@@ -952,7 +953,7 @@
           }
         }
 
-        var match = this.generate(name, params);
+        var match = this.generate(name, params, newContext);
         var entry = this.history[this.history.length - 1];
         var time = entry ? entry.time : getTimeStamp();
 
@@ -1020,6 +1021,14 @@
     }, {
       key: "find",
       value: function find(name, params, newContext) {
+        if (name == undefined) {
+          name = this.name;
+        }
+
+        if (params == undefined) {
+          params = this.params;
+        }
+
         var match = this.generate(name, params, newContext);
         return this.applyFallback(match.url);
       }
@@ -1248,8 +1257,8 @@
         var confirmationEvent = new RelaksRouteManagerEvent('beforechange', this, match);
         var subEntry;
 
-        confirmationEvent.substitute = function (name, params, keepURL) {
-          var sub = _this6.generate(name, params, match.context);
+        confirmationEvent.substitute = function (name, params, newContext, keepURL) {
+          var sub = _this6.generate(name, params, Object.assign({}, match.context, newContext));
 
           if (sub.url === undefined || keepURL) {
             // use URL of the intended route
@@ -1995,22 +2004,28 @@
       this.query = routeManager.query;
       this.search = routeManager.search;
       this.hash = routeManager.hash;
+
+      for (var name in this.route) {
+        if (!this.hasOwnProperty(name)) {
+          this[name] = this.route[name];
+        }
+      }
     }
 
     _createClass(RelaksRouteManagerProxy, [{
       key: "push",
-      value: function push(name, params) {
-        return this.routeManager.push(name, params);
+      value: function push(name, params, context) {
+        return this.routeManager.push(name, params, context);
       }
     }, {
       key: "replace",
-      value: function replace(name, params) {
-        return this.routeManager.replace(name, params);
+      value: function replace(name, params, context) {
+        return this.routeManager.replace(name, params, context);
       }
     }, {
       key: "substitute",
-      value: function substitute(name, params) {
-        return this.routeManager.substitute(name, params);
+      value: function substitute(name, params, context) {
+        return this.routeManager.substitute(name, params, context);
       }
     }, {
       key: "restore",
@@ -2024,8 +2039,8 @@
       }
     }, {
       key: "find",
-      value: function find(name, params) {
-        return this.routeManager.find(name, params);
+      value: function find(name, params, context) {
+        return this.routeManager.find(name, params, context);
       }
     }]);