+++ src/route-manager-proxy.mjs
@@ -12,18 +12,24 @@ class RelaksRouteManagerProxy {
     this.query = routeManager.query;
     this.search = routeManager.search;
     this.hash = routeManager.hash;
+
+    for (let name in this.route) {
+      if (!this.hasOwnProperty(name)) {
+        this[name] = this.route[name];
+      }
+    }
   }
 
-  push(name, params) {
-    return this.routeManager.push(name, params);
+  push(name, params, context) {
+    return this.routeManager.push(name, params, context);
   }
 
-  replace(name, params) {
-    return this.routeManager.replace(name, params);
+  replace(name, params, context) {
+    return this.routeManager.replace(name, params, context);
   }
 
-  substitute(name, params) {
-    return this.routeManager.substitute(name, params);
+  substitute(name, params, context) {
+    return this.routeManager.substitute(name, params, context);
   }
 
   restore() {
@@ -34,8 +40,8 @@ class RelaksRouteManagerProxy {
     return this.routeManager.change(url, options);
   }
 
-  find(name, params) {
-    return this.routeManager.find(name, params);
+  find(name, params, context) {
+    return this.routeManager.find(name, params, context);
   }
 }